buildFrames

+++ R/buildFrameDF.R
@@ -31,6 +31,6 @@ buildFrameDF <- function(df,id,X,Y,domainvalue) {
   stmt <- paste("dframe$domainvalue <- df$",as.character(domainvalue),sep="")
   eval(parse(text=stmt))
   dframe <- as.data.frame(dframe)
-  dframe$domainvalue <- as.numeric(dframe$domainvalue)
+  dframe$domainvalue <- as.integer(dframe$domainvalue)
   return(dframe)
 }
